Move batch generation into ar module
