+++ mlx_vlm/generate/ar.py
@@ -1,24 +1,37 @@
 from __future__ import annotations
 
 import functools
+import contextlib
+import logging
+import os
 import sys
+import time
+import warnings
 from collections.abc import Generator
-from typing import Any, Callable, Dict, List, Optional, Tuple
+from dataclasses import dataclass
+from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 
 import mlx.core as mx
 import mlx.nn as nn
 from mlx_lm.sample_utils import make_logits_processors, make_sampler
 from tqdm import tqdm
 
+from .. import apc as _apc
 from ..models import cache
+from ..prompt_utils import apply_chat_template
 from ..speculative.utils import run_speculative_rounds
+from ..turboquant import BatchTurboQuantKVCache, turboquant_enabled
+from ..utils import group_images_by_shape, prepare_inputs
 from .common import (
     DEFAULT_KV_GROUP_SIZE,
     DEFAULT_KV_QUANT_SCHEME,
     DEFAULT_QUANTIZED_KV_START,
     generation_stream,
     maybe_quantize_kv_cache,
+    wired_limit,
 )
+
+logger = logging.getLogger("mlx_vlm.generate")
 
 DEFAULT_MAX_TOKENS = 2048
 DEFAULT_TEMPERATURE = 0.0
@@ -27,11 +40,25 @@ DEFAULT_TOP_K = 0
 DEFAULT_MIN_P = 0.0
 DEFAULT_REPETITION_CONTEXT_SIZE = 20
 DEFAULT_PREFILL_STEP_SIZE = 2048
+DEFAULT_COMPLETION_BATCH_SIZE = 32
+DEFAULT_PREFILL_BATCH_SIZE = 8
 
 
 def _generate_module_override(name: str, fallback):
     generate_module = sys.modules.get("mlx_vlm.generate")
     return getattr(generate_module, name, fallback) if generate_module else fallback
+
+
+def normalize_resize_shape(values):
+    if values is None:
+        return None
+    if not (
+        not isinstance(values, (str, bytes))
+        and len(values) in (1, 2)
+        and all(type(value) is int for value in values)
+    ):
+        raise ValueError("resize_shape must contain 1 or 2 integers")
+    return (values[0], values[0]) if len(values) == 1 else tuple(values)
 
 
 def generate_step(
@@ -345,3 +372,2108 @@ def generate_step(
         y, logprobs = next_y, next_logprobs
         n += 1
 
+
+@dataclass
+class BatchGenerationResult:
+    """
+    Result of batch generation with optional image size tracking.
+
+    Attributes:
+        texts: Generated text for each sample
+        tokens: Last generated token for each sample
+        logprobs: Log probabilities for each sample
+        prompt_tokens: Number of prompt tokens per sample
+        generation_tokens: Number of generated tokens per sample
+        total_tokens: Total tokens (prompt + generation) per sample
+        prompt_tps: Prompt tokens per second per sample
+        generation_tps: Generation tokens per second per sample
+        peak_memory: Peak memory usage in GB
+        image_sizes: Original (height, width) for each image (for tracking)
+    """
+
+    texts: List[str]
+    tokens: List[Optional[int]]
+    logprobs: List[Optional[List[float]]]
+    prompt_tokens: List[int]
+    generation_tokens: List[int]
+    total_tokens: List[int]
+    prompt_tps: List[float]
+    generation_tps: List[float]
+    peak_memory: float = 0.0
+    image_sizes: Optional[List[Tuple[int, int]]] = None
+
+
+def _left_pad_prompts(prompts, max_length=None):
+    if max_length is None:
+        max_length = max(len(p) for p in prompts)
+
+    return mx.array([[0] * (max_length - len(p)) + p for p in prompts])
+
+
+def _right_pad_prompts(prompts, max_length=None):
+    if max_length is None:
+        max_length = max(len(p) for p in prompts)
+
+    return mx.array([list(p) + [0] * (max_length - len(p)) for p in prompts])
+
+
+_SEQUENCE_ALIGNED_PROMPT_KWARGS = {
+    "attention_mask",
+    "decoder_inputs_embeds",
+    "deepstack_visual_embeds",
+    "visual_pos_masks",
+    "per_layer_inputs",
+    "full_text_row_masked_out_mask",
+    "position_ids",
+    "pos_hw",
+}
+
+APC_PRIVATE_PROMPT_KEYS = ("_apc_tenant", "_apc_image_hash")
+
+
+def _prompt_kwarg_row(v: mx.array, row_idx: int, batch_size: int) -> mx.array:
+    if v.shape[0] == batch_size:
+        return v[row_idx : row_idx + 1]
+    return v[:1]
+
+
+def _split_prompt_kwargs_per_row(prompt_kwargs: dict, batch_size: int) -> List[dict]:
+    """Normalize batched prompt kwargs into one dict per batch row.
+
+    ``model.get_input_embeddings()`` commonly returns batch-sized tensors
+    (notably ``inputs_embeds``). ``BatchGenerator.insert()`` stores prompt
+    kwargs per sequence, so passing the same batched dict for every row causes
+    the prompt builder to concatenate those batched tensors ``batch_size``
+    times, effectively squaring the batch dimension.
+    """
+    if batch_size <= 1:
+        return [prompt_kwargs or {}]
+
+    rows = [{} for _ in range(batch_size)]
+    for k, v in (prompt_kwargs or {}).items():
+        if isinstance(v, mx.array) and v.ndim > 0 and v.shape[0] >= 1:
+            for i in range(batch_size):
+                rows[i][k] = _prompt_kwarg_row(v, i, batch_size)
+        else:
+            for row in rows:
+                row[k] = v
+    return rows
+
+
+def _is_sequence_aligned_prompt_kwarg(
+    key: str, v: mx.array, sequence_length: int
+) -> bool:
+    return (
+        key in _SEQUENCE_ALIGNED_PROMPT_KWARGS
+        and v.ndim >= 2
+        and v.shape[1] == sequence_length
+    )
+
+
+def _pad_sequence_aligned_prompt_kwarg(
+    v: mx.array, target_length: int, *, left: bool
+) -> mx.array:
+    pad = target_length - v.shape[1]
+    if pad <= 0:
+        return v
+    pad_shape = (v.shape[0], pad) + tuple(v.shape[2:])
+    pad_v = mx.zeros(pad_shape, dtype=v.dtype)
+    parts = [pad_v, v] if left else [v, pad_v]
+    return mx.concatenate(parts, axis=1)
+
+
+def _merge_prefill_prompt_kwargs(
+    prompt_kwargs_list: List[Optional[dict]],
+    input_ids: List[List[int]],
+) -> Tuple[mx.array, dict]:
+    """Batch per-row prompt kwargs for a left-padded prefill forward."""
+    lengths = [len(ids) for ids in input_ids]
+    max_length = max(lengths)
+
+    row_embeds: List[mx.array] = []
+    embed_dtype = None
+    embed_dim = None
+    for kw, length in zip(prompt_kwargs_list, lengths):
+        if not kw or kw.get("inputs_embeds") is None:
+            raise ValueError("inputs_embeds is required")
+        embeds = kw["inputs_embeds"]  # [1, length, D]
+        embed_dtype = embeds.dtype
+        embed_dim = embeds.shape[-1]
+        if length < max_length:
+            pad = mx.zeros(
+                (embeds.shape[0], max_length - length, embed_dim),
+                dtype=embed_dtype,
+            )
+            embeds = mx.concatenate([pad, embeds], axis=1)
+        row_embeds.append(embeds)
+    inputs_embeds = mx.concatenate(row_embeds, axis=0)
+
+    merged_kwargs: dict = {}
+    per_row_keys: dict = {}
+    batch_size = len(prompt_kwargs_list)
+    for i, (kw, length) in enumerate(zip(prompt_kwargs_list, lengths)):
+        if not kw:
+            continue
+        for k, v in kw.items():
+            if k == "inputs_embeds" or k in APC_PRIVATE_PROMPT_KEYS:
+                continue
+            if isinstance(v, mx.array) and v.ndim > 0 and v.shape[0] >= 1:
+                row_v = _prompt_kwarg_row(v, i, batch_size)
+                if _is_sequence_aligned_prompt_kwarg(k, row_v, length):
+                    row_v = _pad_sequence_aligned_prompt_kwarg(
+                        row_v, max_length, left=True
+                    )
+                per_row_keys.setdefault(k, []).append(row_v)
+            elif k not in merged_kwargs:
+                merged_kwargs[k] = v
+    for k, vs in per_row_keys.items():
+        merged_kwargs[k] = mx.concatenate(vs, axis=0)
+
+    return inputs_embeds, merged_kwargs
+
+
+def _extend_cache(cache_a, cache_b):
+    """Extend cache_a with cache_b along the batch dimension."""
+    if not cache_a:
+        return cache_b
+    if not cache_b:
+        return cache_a
+    for ca, cb in zip(cache_a, cache_b):
+        ca.extend(cb)
+    return cache_a
+
+
+def _make_cache(
+    model,
+    left_padding,
+    kv_bits=None,
+    kv_group_size=64,
+    kv_quant_scheme=DEFAULT_KV_QUANT_SCHEME,
+):
+    """
+    Convert a list of regular caches into their corresponding
+    batch-aware caches.
+
+    When *kv_bits* is set, a quantized batch cache is used instead of
+    ``BatchKVCache`` so that KV states are quantized on-the-fly during
+    generation, reducing memory usage for long sequences.
+
+    *kv_quant_scheme* selects the quantization backend:
+    - ``"uniform"`` → ``BatchQuantizedKVCache`` (``mx.quantize``)
+    - ``"turboquant"`` or fractional *kv_bits* → ``BatchTurboQuantKVCache``
+    """
+    use_turbo = kv_bits is not None and turboquant_enabled(kv_bits, kv_quant_scheme)
+
+    def _make_quant_cache(lp):
+        if use_turbo:
+            return BatchTurboQuantKVCache(lp, bits=kv_bits)
+        return cache.BatchQuantizedKVCache(
+            lp, group_size=kv_group_size, bits=int(kv_bits)
+        )
+
+    def to_batch_cache(c, quantize=True):
+        if isinstance(c, cache.KVCache):
+            if kv_bits is not None and quantize:
+                return _make_quant_cache(left_padding)
+            return cache.BatchKVCache(left_padding)
+        elif isinstance(c, cache.ChunkedKVCache):
+            if kv_bits is not None and quantize:
+                return _make_quant_cache(left_padding)
+            return cache.BatchKVCache(left_padding)
+        elif isinstance(c, cache.SimpleKVCache):
+            if kv_bits is not None and quantize:
+                return _make_quant_cache(left_padding)
+            return cache.BatchKVCache(left_padding)
+        elif isinstance(c, cache.ArraysCache):
+            c.left_padding = mx.array(left_padding)
+            return c
+        elif isinstance(c, cache.RotatingKVCache):
+            if c.keep > 0:
+                raise ValueError("RotatingKVCache with keep tokens is not supported.")
+            return cache.BatchRotatingKVCache(c.max_size, left_padding)
+        elif isinstance(c, cache.CacheList):
+            return cache.CacheList(*(to_batch_cache(sub_c) for sub_c in c.caches))
+        elif isinstance(c, tuple):
+            return cache.CacheList(*(to_batch_cache(sub_c) for sub_c in c))
+        else:
+            raise ValueError(f"{type(c)} does not yet support batching")
+
+    if hasattr(model, "make_cache"):
+        model_cache = model.make_cache()
+        n = len(model_cache)
+        # Skip quantizing the last layer — it's sensitive to quantization
+        return [
+            to_batch_cache(c, quantize=(i < n - 1 if n > 2 else True))
+            for i, c in enumerate(model_cache)
+        ]
+    else:
+        if kv_bits is not None:
+            n = len(model.layers)
+            return [
+                (
+                    _make_quant_cache(left_padding)
+                    if i < n - 1 or n <= 2
+                    else cache.BatchKVCache(left_padding)
+                )
+                for i in range(n)
+            ]
+        return [cache.BatchKVCache(left_padding) for _ in model.layers]
+
+
+@dataclass
+class BatchStats:
+    """
+    An data object to hold generation stats.
+
+    Args:
+        prompt_tokens (int): The number of prompt tokens processed.
+        prompt_tps (float): The prompt processing tokens-per-second.
+        prompt_time (float): The time in seconds spent in prompt processing.
+        generation_tokens (int): The number of generated tokens.
+        generation_tps (float): The tokens-per-second for generation.
+        generation_time (float): The time in seconds spent in generation .
+        peak_memory (float): The peak memory used so far in GB.
+    """
+
+    prompt_tokens: int = 0
+    prompt_tps: float = 0
+    prompt_time: float = 0
+    generation_tokens: int = 0
+    generation_tps: float = 0
+    generation_time: float = 0
+    peak_memory: float = 0
+
+
+@dataclass
+class BatchResponse:
+    """
+    An data object to hold a batch generation response.
+
+    Args:
+        texts: (List[str]): The generated text for each prompt.
+        stats (BatchStats): Statistics about the generation.
+        image_sizes: (Optional[List[Tuple[int, int]]]): Original (height, width)
+            for each image. Useful for tracking which images produced which responses
+            and for debugging padding/batching behavior.
+    """
+
+    texts: List[str]
+    stats: BatchStats
+    image_sizes: Optional[List[Tuple[int, int]]] = None
+
+
+@dataclass
+class PromptProgress:
+    """Per-request prompt processing metrics for continuous batching."""
+
+    uid: int
+    prompt_tokens: int
+    prompt_tps: float = 0.0
+    prompt_time: float = 0.0
+    cached_tokens: int = 0
+
+
+class GenerationBatch:
+    """
+    Batched token generator with double-buffered pipelining.
+
+    Manages the generation phase after prompt processing, with KV caches,
+    sampling, and stop detection for multiple sequences. Uses async_eval
+    to overlap GPU computation with CPU processing (decode-ahead pattern).
+    """
+
+    @dataclass
+    class Response:
+        uid: int
+        token: int
+        token_logprob: float
+        finish_reason: Optional[str]
+        top_logprobs: Optional[List[Tuple[int, float]]] = None
+
+    def __init__(
+        self,
+        model: nn.Module,
+        uids: List[int],
+        inputs: mx.array,
+        prompt_cache: List[Any],
+        sampler: Callable[[mx.array], mx.array],
+        stop_criteria,
+        max_tokens: List[int],
+        top_logprobs_k: int = 0,
+        token_context: Optional[List[List[int]]] = None,
+        logits_processors: Optional[
+            List[Optional[List[Callable[[mx.array, mx.array], mx.array]]]]
+        ] = None,
+    ):
+        self.model = model
+        self._language_model = getattr(model, "language_model", model)
+        self.uids = uids
+        self.prompt_cache = prompt_cache
+        self.sampler = sampler
+        self.stop_criteria = stop_criteria
+        self.max_tokens = max_tokens
+        self._num_tokens = [0] * len(uids)
+        self.compute_logprobs = True
+        self.top_logprobs_k = top_logprobs_k
+        self.logits_processors = logits_processors or []
+        self.token_context = [list(ctx) for ctx in (token_context or [])]
+        self._ensure_token_context()
+
+        self._current_tokens = None
+        self._current_lps = None
+        self._next_tokens = inputs
+        self._next_lps = None
+        self._next_top_idx = None
+        self._next_top_lp = None
+
+        # Per-sequence MRoPE delta
+        self._rope_deltas = None
+
+    def __len__(self):
+        return len(self.uids)
+
+    def _ensure_token_context(self, *, force: bool = False):
+        if not (force or (self.logits_processors and any(self.logits_processors))):
+            if not self.logits_processors:
+                self.token_context = []
+            return
+        if len(self.token_context) < len(self.uids):
+            missing = len(self.uids) - len(self.token_context)
+            self.token_context.extend([[] for _ in range(missing)])
+        elif len(self.token_context) > len(self.uids):
+            self.token_context = self.token_context[: len(self.uids)]
+
+    def _step(self):
+        """Perform one generation step with double buffering."""
+        self._current_tokens = self._next_tokens
+        self._current_lps = self._next_lps
+        inputs = self._current_tokens
+
+        fwd_kwargs = {}
+        if self._rope_deltas is not None:
+            fwd_kwargs["rope_deltas"] = self._rope_deltas
+
+        output = self._language_model(
+            inputs[:, None], cache=self.prompt_cache, **fwd_kwargs
+        )
+        logits = output.logits if hasattr(output, "logits") else output
+        logits = logits[:, -1, :]
+
+        if self.logits_processors and any(self.logits_processors):
+            last_tokens = inputs.tolist()
+            self._ensure_token_context()
+            for i, token in enumerate(last_tokens):
+                self.token_context[i].append(token)
+
+            processed_logits = []
+            for i in range(logits.shape[0]):
+                sample_logits = logits[i : i + 1]
+                processors = self.logits_processors[i] or []
+                for processor in processors:
+                    if hasattr(processor, "process_last_token"):
+                        sample_logits = processor.process_last_token(
+                            last_tokens[i], sample_logits
+                        )
+                    else:
+                        sample_logits = processor(
+                            mx.array(self.token_context[i]), sample_logits
+                        )
+                processed_logits.append(sample_logits)
+            logits = mx.concatenate(processed_logits, axis=0)
+
+        logprobs = logits - mx.logsumexp(logits, axis=-1, keepdims=True)
+        sampled = self.sampler(logprobs)
+
+        self._next_tokens = sampled
+        prev_top_idx = self._next_top_idx
+        prev_top_lp = self._next_top_lp
+
+        eval_targets = [self._next_tokens]
+        if self.compute_logprobs:
+            self._next_lps = logprobs[mx.arange(sampled.shape[0]), sampled]
+            eval_targets.append(self._next_lps)
+        else:
+            self._next_lps = None
+
+        k = self.top_logprobs_k
+        if k > 0:
+            # argsort ascending; take last K columns and reverse for descending.
+            sort_idx = mx.argsort(logprobs, axis=-1)
+            top_idx = sort_idx[..., -k:][..., ::-1].astype(mx.int32)
+            top_lp = mx.take_along_axis(logprobs, top_idx, axis=-1)
+            self._next_top_idx = top_idx
+            self._next_top_lp = top_lp
+            eval_targets.extend([top_idx, top_lp])
+        else:
+            self._next_top_idx = None
+            self._next_top_lp = None
+
+        mx.async_eval(*eval_targets)
+
+        if self._current_lps is not None:
+            to_eval = [inputs, self._current_lps]
+            if prev_top_idx is not None:
+                to_eval.extend([prev_top_idx, prev_top_lp])
+            mx.eval(*to_eval)
+            top_idx_list = prev_top_idx.tolist() if prev_top_idx is not None else None
+            top_lp_list = prev_top_lp.tolist() if prev_top_lp is not None else None
+            return (
+                inputs.tolist(),
+                self._current_lps.tolist(),
+                top_idx_list,
+                top_lp_list,
+            )
+        else:
+            mx.eval(inputs)
+            return inputs.tolist(), None, None, None
+
+    def _eval_pending_state(self):
+        """Materialize lazy decode outputs before mutating batch-owned state."""
+        targets = []
+
+        def append_arrays(value):
+            if isinstance(value, mx.array):
+                targets.append(value)
+            elif isinstance(value, (list, tuple)):
+                for item in value:
+                    append_arrays(item)
+
+        append_arrays(
+            (
+                self._current_tokens,
+                self._current_lps,
+                self._next_tokens,
+                self._next_lps,
+                self._next_top_idx,
+                self._next_top_lp,
+                self._rope_deltas,
+            )
+        )
+        for c in self.prompt_cache:
+            try:
+                append_arrays(c.state)
+            except (AttributeError, TypeError):
+                pass
+
+        if targets:
+            mx.eval(*targets)
+
+    def extend(self, other: "GenerationBatch"):
+        """Extend this batch with another generation batch."""
+        self_was_empty = len(self.uids) == 0
+        if not self_was_empty and len(other.uids) > 0:
+            self._eval_pending_state()
+            other._eval_pending_state()
+
+        self_has_processors = self.logits_processors and any(self.logits_processors)
+        other_has_processors = other.logits_processors and any(other.logits_processors)
+        if self_has_processors or other_has_processors:
+            self._ensure_token_context(force=bool(other_has_processors))
+            other._ensure_token_context(force=bool(self_has_processors))
+
+        self.uids.extend(other.uids)
+        self.prompt_cache = _extend_cache(self.prompt_cache, other.prompt_cache)
+        self.max_tokens.extend(other.max_tokens)
+        self._num_tokens.extend(other._num_tokens)
+        self.token_context.extend(other.token_context)
+        self.logits_processors.extend(other.logits_processors)
+        self._ensure_token_context()
+
+        if self._current_tokens is None:
+            self._current_tokens = other._current_tokens
+            self._current_lps = other._current_lps
+        elif other._current_tokens is not None:
+            self._current_tokens = mx.concatenate(
+                [self._current_tokens, other._current_tokens]
+            )
+            if self._current_lps is not None and other._current_lps is not None:
+                self._current_lps = mx.concatenate(
+                    [self._current_lps, other._current_lps]
+                )
+
+        if self._next_tokens is None:
+            self._next_tokens = other._next_tokens
+            self._next_lps = other._next_lps
+            self._next_top_idx = other._next_top_idx
+            self._next_top_lp = other._next_top_lp
+        elif other._next_tokens is not None:
+            self._next_tokens = mx.concatenate([self._next_tokens, other._next_tokens])
+            if self._next_lps is not None and other._next_lps is not None:
+                self._next_lps = mx.concatenate([self._next_lps, other._next_lps])
+
+            if (
+                self._next_top_idx is not None
+                and other._next_top_idx is not None
+                and self._next_top_idx.shape[-1] == other._next_top_idx.shape[-1]
+            ):
+                self._next_top_idx = mx.concatenate(
+                    [self._next_top_idx, other._next_top_idx]
+                )
+                self._next_top_lp = mx.concatenate(
+                    [self._next_top_lp, other._next_top_lp]
+                )
+            else:
+                self._next_top_idx = None
+                self._next_top_lp = None
+
+        if self_was_empty:
+            self._rope_deltas = other._rope_deltas
+        elif (self._rope_deltas is None) != (other._rope_deltas is None):
+            raise RuntimeError(
+                "extend() mixes MRoPE and non-MRoPE batches; both sides must "
+                "carry rope_deltas or neither side may."
+            )
+        elif self._rope_deltas is not None:
+            self._rope_deltas = mx.concatenate([self._rope_deltas, other._rope_deltas])
+
+    def filter(self, keep: List[int]):
+        """Filter the batch to keep only the specified indices."""
+        if len(keep) < len(self.uids):
+            self._eval_pending_state()
+
+        self.uids = [self.uids[idx] for idx in keep]
+        self.max_tokens = [self.max_tokens[idx] for idx in keep]
+        self._num_tokens = [self._num_tokens[idx] for idx in keep]
+        if self.token_context:
+            self.token_context = [self.token_context[idx] for idx in keep]
+        if self.logits_processors:
+            self.logits_processors = [self.logits_processors[idx] for idx in keep]
+
+        if not keep:
+            self.prompt_cache.clear()
+            self._current_tokens = None
+            self._current_lps = None
+            self._next_tokens = None
+            self._next_lps = None
+            self._next_top_idx = None
+            self._next_top_lp = None
+            self._rope_deltas = None
+            self.token_context = []
+            self.logits_processors = []
+        else:
+            keep_arr = mx.array(keep, mx.int32)
+            for c in self.prompt_cache:
+                c.filter(keep_arr)
+            if self._next_tokens is not None:
+                self._next_tokens = self._next_tokens[keep_arr]
+            if self._next_lps is not None:
+                self._next_lps = self._next_lps[keep_arr]
+            if self._next_top_idx is not None:
+                self._next_top_idx = self._next_top_idx[keep_arr]
+                self._next_top_lp = self._next_top_lp[keep_arr]
+            if self._rope_deltas is not None:
+                self._rope_deltas = self._rope_deltas[keep_arr]
+
+    def next(self) -> List[Response]:
+        """Generate the next batch of tokens."""
+        if not self.uids:
+            return []
+
+        tokens, lp_list, top_idx_list, top_lp_list = self._step()
+
+        keep = []
+        responses = []
+        for i in range(len(self.uids)):
+            finish_reason = None
+            self._num_tokens[i] += 1
+            tok = tokens[i]
+
+            if self.stop_criteria(tok):
+                finish_reason = "stop"
+            elif self._num_tokens[i] >= self.max_tokens[i]:
+                finish_reason = "length"
+
+            if finish_reason is None:
+                keep.append(i)
+
+            top_lp = None
+            if top_idx_list is not None:
+                top_lp = list(zip(top_idx_list[i], top_lp_list[i]))
+
+            responses.append(
+                self.Response(
+                    uid=self.uids[i],
+                    token=tok,
+                    token_logprob=lp_list[i] if lp_list is not None else 0.0,
+                    finish_reason=finish_reason,
+                    top_logprobs=top_lp,
+                )
+            )
+
+        if len(keep) < len(self.uids):
+            self.filter(keep)
+
+        return responses
+
+    @classmethod
+    def empty(
+        cls, model, sampler, stop_criteria, compute_logprobs=True, top_logprobs_k=0
+    ):
+        """Create an empty generation batch."""
+        batch = cls.__new__(cls)
+        batch.model = model
+        batch._language_model = getattr(model, "language_model", model)
+        batch.uids = []
+        batch.prompt_cache = []
+        batch.sampler = sampler
+        batch.stop_criteria = stop_criteria
+        batch.max_tokens = []
+        batch._num_tokens = []
+        batch.compute_logprobs = compute_logprobs
+        batch.top_logprobs_k = top_logprobs_k
+        batch.token_context = []
+        batch.logits_processors = []
+        batch._current_tokens = None
+        batch._current_lps = None
+        batch._next_tokens = None
+        batch._next_lps = None
+        batch._next_top_idx = None
+        batch._next_top_lp = None
+        batch._rope_deltas = None
+        return batch
+
+
+class PromptProcessingBatch:
+    """
+    Handles VLM prompt processing with inputs_embeds and chunked prefill.
+
+    Processes prompt tokens incrementally (one chunk per step) to allow
+    interleaving with generation for continuous batching. Transitions to
+    a GenerationBatch when prompt processing is complete.
+    """
+
+    def __init__(
+        self,
+        model: nn.Module,
+        uids: List[int],
+        input_ids: List[List[int]],
+        max_tokens: List[int],
+        inputs_embeds: mx.array,
+        prompt_kwargs: dict,
+        logits_processors: Optional[
+            List[Optional[List[Callable[[mx.array, mx.array], mx.array]]]]
+        ] = None,
+        prefill_step_size: Optional[int] = DEFAULT_PREFILL_STEP_SIZE,
+        kv_bits=None,
+        kv_group_size: int = DEFAULT_KV_GROUP_SIZE,
+        kv_quant_scheme: str = DEFAULT_KV_QUANT_SCHEME,
+        warm_cache: Optional[List[Any]] = None,
+        apc_meta: Optional[List[dict]] = None,
+        apc_manager: Optional["_apc.APCManager"] = None,
+        right_pad_per_row: Optional[List[int]] = None,
+        suffix_lens: Optional[List[int]] = None,
+        apc_mode: Optional[str] = None,
+    ):
+        self.model = model
+        self.uids = uids
+        self._prompt_uids = list(uids)
+        self.max_tokens = max_tokens
+        self.prefill_step_size = prefill_step_size
+
+        lengths = [len(ids) for ids in input_ids]
+        max_length = max(lengths)
+        # ``input_ids`` here are the per-row prefill inputs — for warm-start
+        # rows this is the suffix, for cold rows the full prompt. When
+        # ``right_pad_per_row`` is set the rows are right-padded (used in
+        # mixed warm/cold prefill so suffix RoPE positions align). Otherwise
+        # we left-pad as before.
+        self._right_pad_per_row = right_pad_per_row
+        self._suffix_lens = suffix_lens or lengths
+        self._left_padding_per_row: List[int]
+
+        if right_pad_per_row is not None:
+            # Right-pad each row to max_length (so the last `pad[i]` cells are
+            # right-pad and need to be rolled into left-pad by finalize()).
+            left_padding = [0] * len(input_ids)
+            self._input_ids = _right_pad_prompts(input_ids, max_length=max_length)
+        else:
+            left_padding = [max_length - l for l in lengths]
+            self._input_ids = _left_pad_prompts(input_ids, max_length=max_length)
+        self._left_padding_per_row = list(left_padding)
+        self._total_prompt_tokens = sum(lengths)
+        self._processed_prompt_columns = 0
+
+        self.logits_processors = logits_processors or []
+        self._token_context = (
+            [list(ids) for ids in input_ids]
+            if self.logits_processors and any(self.logits_processors)
+            else []
+        )
+        self._inputs_embeds = inputs_embeds
+        self._prompt_kwargs = prompt_kwargs or {}
+        self._prompt_length_aware_keys: List[str] = []
+        if self._prompt_kwargs and self._inputs_embeds is not None:
+            prompt_batch = self._inputs_embeds.shape[0]
+            prompt_len = self._inputs_embeds.shape[1]
+            for k, v in self._prompt_kwargs.items():
+                if (
+                    isinstance(v, mx.array)
+                    and v.ndim >= 2
+                    and v.shape[0] == prompt_batch
+                    and v.shape[1] == prompt_len
+                ):
+                    self._prompt_length_aware_keys.append(k)
+
+        # APC metadata used for post-prefill block harvest (per-row).
+        self._apc_meta = apc_meta or []
+        self._apc_manager = apc_manager
+        self._apc_mode = apc_mode
+        self._apc_harvest_enabled = True
+        self._prompt_time_s = 0.0
+        self._prompt_tokens_per_row: List[int] = []
+        self._cached_tokens_per_row: List[int] = []
+        for idx, suffix_len in enumerate(lengths):
+            full_input_ids = None
+            prefix_len = 0
+            if idx < len(self._apc_meta) and self._apc_meta[idx] is not None:
+                full_input_ids = self._apc_meta[idx].get("full_input_ids")
+                prefix_len = int(self._apc_meta[idx].get("prefix_len") or 0)
+            self._prompt_tokens_per_row.append(
+                len(full_input_ids) if full_input_ids is not None else suffix_len
+            )
+            self._cached_tokens_per_row.append(prefix_len)
+
+        if warm_cache is not None:
+            self.prompt_cache = warm_cache
+        else:
+            self.prompt_cache = _make_cache(
+                model,
+                left_padding,
+                kv_bits=kv_bits,
+                kv_group_size=kv_group_size,
+                kv_quant_scheme=kv_quant_scheme,
+            )
+
+        # Declare per-row right-padding on each cache so finalize() can roll
+        # it into left-padding once the prefill forward pass is complete.
+        if right_pad_per_row is not None and any(right_pad_per_row):
+            for c in self.prompt_cache:
+                prepare = getattr(c, "prepare", None)
+                if not callable(prepare):
+                    self._apc_harvest_enabled = False
+                    self._release_apc_meta_blocks()
+                    raise RuntimeError(
+                        "APC mixed prefill requires a prompt cache with prepare()"
+                    )
+                prepare(right_padding=right_pad_per_row, lengths=self._suffix_lens)
+
+    def __len__(self):
+        return len(self.uids)
+
+    def _release_apc_meta_blocks(self):
+        if self._apc_manager is None:
+            return
+        for meta in self._apc_meta:
+            if meta is not None:
+                self._apc_manager.release(meta.get("apc_blocks", []))
+
+    def needs_processing(self):
+        """True if prompt needs chunked processing before generate()."""
+        if self._inputs_embeds is None or self.prefill_step_size is None:
+            return self._next_apc_checkpoint_column() is not None
+        if self._next_apc_checkpoint_column() is not None:
+            return True
+        return self._inputs_embeds.shape[1] > self.prefill_step_size
+
+    def _apc_checkpoint_column_for_meta(
+        self, batch_idx: int, meta: dict
+    ) -> Optional[int]:
+        checkpoint_len = int(meta.get("checkpoint_len") or 0)
+        if (
+            self._apc_mode != "exact"
+            or checkpoint_len <= 0
+            or meta.get("checkpoint_done")
+        ):
+            return None
+        prefix_len = int(meta.get("prefix_len", 0) or 0)
+        if checkpoint_len <= prefix_len:
+            meta["checkpoint_done"] = True
+            return None
+        if self._right_pad_per_row is not None:
+            suffix_checkpoint = checkpoint_len - prefix_len
+            if suffix_checkpoint >= self._suffix_lens[batch_idx]:
+                return None
+            return suffix_checkpoint
+        return self._left_padding_per_row[batch_idx] + checkpoint_len
+
+    def _next_apc_checkpoint_column(self) -> Optional[int]:
+        if (
+            self._apc_manager is None
+            or self._apc_mode != "exact"
+            or not self._apc_meta
+            or self._inputs_embeds is None
+        ):
+            return None
+        start = self._processed_prompt_columns
+        end = start + self._inputs_embeds.shape[1]
+        next_col: Optional[int] = None
+        for batch_idx, meta in enumerate(self._apc_meta):
+            if meta is None:
+                continue
+            col = self._apc_checkpoint_column_for_meta(batch_idx, meta)
+            if col is None or col <= start or col >= end:
+                continue
+            next_col = col if next_col is None else min(next_col, col)
+        return next_col
+
+    def _row_real_tokens_processed(self, batch_idx: int) -> int:
+        meta = self._apc_meta[batch_idx]
+        prefix_len = int(meta.get("prefix_len", 0) or 0)
+        if self._right_pad_per_row is not None:
+            suffix_done = min(
+                self._suffix_lens[batch_idx],
+                max(0, self._processed_prompt_columns),
+            )
+            return prefix_len + suffix_done
+        real_done = (
+            self._processed_prompt_columns - self._left_padding_per_row[batch_idx]
+        )
+        return prefix_len + min(self._suffix_lens[batch_idx], max(0, real_done))
+
+    def _store_apc_exact_checkpoints(self) -> None:
+        if self._apc_manager is None or self._apc_mode != "exact":
+            return
+        for batch_idx, meta in enumerate(self._apc_meta):
+            if meta is None or meta.get("checkpoint_done"):
+                continue
+            checkpoint_len = int(meta.get("checkpoint_len") or 0)
+            if checkpoint_len <= 0:
+                continue
+            if self._row_real_tokens_processed(batch_idx) != checkpoint_len:
+                continue
+            prompt_cache = _apc.extract_prompt_cache_from_batch(
+                self.prompt_cache,
+                batch_idx,
+            )
+            if prompt_cache is None:
+                continue
+            self._apc_manager.store_exact_cache(
+                meta["full_input_ids"][:checkpoint_len],
+                prompt_cache,
+                extra_hash=meta.get("extra_hash", 0),
+            )
+            meta["checkpoint_done"] = True
+
+    def _prompt_kwargs_for_step(self, n: Optional[int] = None) -> dict:
+        if n is None or not self._prompt_length_aware_keys:
+            return self._prompt_kwargs
+        out = dict(self._prompt_kwargs)
+        for k in self._prompt_length_aware_keys:
+            out[k] = out[k][:, :n, ...]
+        return out
+
+    def prompt_step(self) -> int:
+        """Process one chunk of the prompt. Returns tokens processed."""
+        if not self.needs_processing():
+            return 0
+
+        step = self.prefill_step_size or self._inputs_embeds.shape[1]
+        n = min(step, self._inputs_embeds.shape[1] - 1)
+        checkpoint_col = self._next_apc_checkpoint_column()
+        if checkpoint_col is not None:
+            n = min(n, checkpoint_col - self._processed_prompt_columns)
+        if n <= 0:
+            return 0
+        prompt_kwargs = self._prompt_kwargs_for_step(n)
+        self.model(
+            self._input_ids[:, :n],
+            cache=self.prompt_cache,
+            inputs_embeds=self._inputs_embeds[:, :n],
+            n_to_process=n,
+            **prompt_kwargs,
+        )
+        mx.eval([c.state for c in self.prompt_cache])
+        self._processed_prompt_columns += n
+        self._store_apc_exact_checkpoints()
+        self._inputs_embeds = self._inputs_embeds[:, n:]
+        self._input_ids = self._input_ids[:, n:]
+        for k in self._prompt_length_aware_keys:
+            self._prompt_kwargs[k] = self._prompt_kwargs[k][:, n:, ...]
+        mx.clear_cache()
+        return n
+
+    def record_prompt_time(self, elapsed_s: float) -> None:
+        self._prompt_time_s += max(0.0, float(elapsed_s))
+
+    def prompt_progress(self) -> List[PromptProgress]:
+        if self._prompt_time_s <= 0:
+            return []
+        return [
+            PromptProgress(
+                uid=uid,
+                prompt_tokens=prompt_tokens,
+                prompt_tps=prompt_tokens / self._prompt_time_s,
+                prompt_time=self._prompt_time_s,
+                cached_tokens=cached_tokens,
+            )
+            for uid, prompt_tokens, cached_tokens in zip(
+                self._prompt_uids,
+                self._prompt_tokens_per_row,
+                self._cached_tokens_per_row,
+            )
+        ]
+
+    def generate(
+        self, sampler, stop_criteria, compute_logprobs=True, top_logprobs_k=0
+    ) -> GenerationBatch:
+        """Process final tokens and transition to GenerationBatch."""
+        output = self.model(
+            self._input_ids,
+            cache=self.prompt_cache,
+            inputs_embeds=self._inputs_embeds,
+            **self._prompt_kwargs,
+        )
+        logits = output.logits if hasattr(output, "logits") else output
+        if self._right_pad_per_row is not None and any(self._right_pad_per_row):
+            # Per-row last *real* token sits at index (seq - 1 - right_pad[i]).
+            seq = logits.shape[1]
+            last_idx = mx.array(
+                [seq - 1 - p for p in self._right_pad_per_row], dtype=mx.int32
+            )[:, None, None]
+            last_idx = mx.broadcast_to(last_idx, (logits.shape[0], 1, logits.shape[-1]))
+            logits = mx.take_along_axis(logits, last_idx, axis=1).squeeze(1)
+        else:
+            logits = logits[:, -1, :]
+        if self.logits_processors and any(self.logits_processors):
+            processed_logits = []
+            for i in range(logits.shape[0]):
+                sample_logits = logits[i : i + 1]
+                processors = self.logits_processors[i] or []
+                for processor in processors:
+                    sample_logits = processor(
+                        mx.array(self._token_context[i]), sample_logits
+                    )
+                processed_logits.append(sample_logits)
+            logits = mx.concatenate(processed_logits, axis=0)
+
+        logprobs = logits - mx.logsumexp(logits, axis=-1, keepdims=True)
+        first_tokens = sampler(logprobs)
+
+        mx.async_eval(first_tokens)
+
+        # Roll any right-padding into left-padding so the cache decoded by
+        # GenerationBatch sees a canonical layout.
+        if self._right_pad_per_row is not None and any(self._right_pad_per_row):
+            for c in self.prompt_cache:
+                finalize = getattr(c, "finalize", None)
+                if not callable(finalize):
+                    self._apc_harvest_enabled = False
+                    self._release_apc_meta_blocks()
+                    raise RuntimeError(
+                        "APC mixed prefill requires a prompt cache with finalize()"
+                    )
+                finalize()
+        if logger.isEnabledFor(logging.DEBUG) and os.environ.get("APC_DEBUG"):
+            c0 = self.prompt_cache[0] if self.prompt_cache else None
+            if c0 is not None:
+                off = getattr(c0, "offset", None)
+                lp = getattr(c0, "left_padding", None)
+                logger.warning(
+                    "post-prefill cache[0]: _idx=%s offset=%s left_padding=%s right_pad_per_row=%s suffix_lens=%s",
+                    getattr(c0, "_idx", None),
+                    off.tolist() if hasattr(off, "tolist") else off,
+                    lp.tolist() if hasattr(lp, "tolist") else lp,
+                    self._right_pad_per_row,
+                    self._suffix_lens,
+                )
+
+        gen_batch = GenerationBatch(
+            model=self.model,
+            uids=list(self.uids),
+            inputs=first_tokens,
+            prompt_cache=self.prompt_cache,
+            sampler=sampler,
+            stop_criteria=stop_criteria,
+            max_tokens=list(self.max_tokens),
+            top_logprobs_k=top_logprobs_k,
+            token_context=[list(ctx) for ctx in self._token_context],
+            logits_processors=list(self.logits_processors),
+        )
+        gen_batch.compute_logprobs = compute_logprobs
+
+        if compute_logprobs:
+            gen_batch._next_lps = logprobs[
+                mx.arange(first_tokens.shape[0]), first_tokens
+            ]
+
+        # Prime top-K buffers so the first token can emit top_logprobs too.
+        if top_logprobs_k > 0:
+            k = top_logprobs_k
+            sort_idx = mx.argsort(logprobs, axis=-1)
+            top_idx = sort_idx[..., -k:][..., ::-1].astype(mx.int32)
+            top_lp = mx.take_along_axis(logprobs, top_idx, axis=-1)
+            gen_batch._next_top_idx = top_idx
+            gen_batch._next_top_lp = top_lp
+
+        language_model = getattr(self.model, "language_model", self.model)
+        rope_deltas = self._capture_rope_deltas(language_model, len(gen_batch.uids))
+        if rope_deltas is not None:
+            # Normalize to shape (B, 1) so extend/filter stay consistent.
+            if rope_deltas.ndim == 0:
+                rope_deltas = rope_deltas.reshape(1, 1)
+            elif rope_deltas.ndim == 1:
+                rope_deltas = rope_deltas[:, None]
+            # When a warm-start batch reuses the model's cached _rope_deltas
+            # (computed during a previous prefill with a smaller batch), the
+            # batch dim won't match this prompt batch's row count. Realign
+            # so extend()/filter() down the line stay consistent with the
+            # generation batch's row count.
+            target_b = first_tokens.shape[0]
+            if rope_deltas.shape[0] != target_b:
+                if rope_deltas.shape[0] == 1:
+                    rope_deltas = mx.broadcast_to(
+                        rope_deltas, (target_b, rope_deltas.shape[1])
+                    )
+                elif rope_deltas.shape[0] < target_b:
+                    pad = target_b - rope_deltas.shape[0]
+                    rope_deltas = mx.concatenate(
+                        [
+                            rope_deltas,
+                            mx.broadcast_to(
+                                rope_deltas[-1:],
+                                (pad, rope_deltas.shape[1]),
+                            ),
+                        ],
+                        axis=0,
+                    )
+                else:
+                    rope_deltas = rope_deltas[:target_b]
+            gen_batch._rope_deltas = rope_deltas
+
+        # APC: harvest the post-prefill K/V into hashed blocks. Done after the
+        # final prefill forward but before the cache references are released
+        # so the block tensors snapshot the prompt prefix.
+        if (
+            self._apc_manager is not None
+            and self._apc_meta
+            and self._apc_harvest_enabled
+        ):
+            try:
+                for batch_idx, meta in enumerate(self._apc_meta):
+                    if meta is None:
+                        continue
+                    if self._apc_mode == "exact":
+                        prompt_cache = _apc.extract_prompt_cache_from_batch(
+                            self.prompt_cache,
+                            batch_idx,
+                        )
+                        if prompt_cache is not None:
+                            self._apc_manager.store_exact_cache(
+                                meta["full_input_ids"],
+                                prompt_cache,
+                                extra_hash=meta.get("extra_hash", 0),
+                            )
+                        self._apc_manager.release(meta.get("apc_blocks", []))
+                    else:
+                        new_blocks = _apc.harvest_blocks_from_batch_cache(
+                            self._apc_manager,
+                            self.prompt_cache,
+                            batch_idx,
+                            meta["full_input_ids"],
+                            extra_hash=meta.get("extra_hash", 0),
+                            skip_first_n_tokens=meta.get("prefix_len", 0),
+                        )
+                        self._apc_manager.release(
+                            meta.get("apc_blocks", []) + new_blocks
+                        )
+            except Exception as e:
+                logger.warning("APC harvest failed during batched prefill: %s", e)
+                # Best effort — release any acquired prefix blocks.
+                for meta in self._apc_meta:
+                    if meta is not None:
+                        self._apc_manager.release(meta.get("apc_blocks", []))
+
+        self.uids = []
+        self.prompt_cache = []
+        self._token_context = []
+        self.logits_processors = []
+        self._apc_meta = []
+        return gen_batch
+
+    @property
+    def total_prompt_tokens(self):
+        return self._total_prompt_tokens
+
+    @staticmethod
+    def _capture_rope_deltas(language_model, B: int):
+        if not hasattr(language_model, "_rope_deltas"):
+            return None
+        rope_deltas = language_model._rope_deltas
+        if rope_deltas is None:
+            return mx.zeros((B, 1), dtype=mx.int32)
+        if rope_deltas.ndim == 0:
+            rope_deltas = rope_deltas.reshape(1, 1)
+        elif rope_deltas.ndim == 1:
+            rope_deltas = rope_deltas[:, None]
+        # Falcon OCR emits a singleton meant to broadcast across rows.
+        if rope_deltas.shape[0] == 1 and B > 1:
+            rope_deltas = mx.broadcast_to(rope_deltas, (B, 1))
+        if rope_deltas.shape[0] != B:
+            raise RuntimeError(
+                f"_rope_deltas shape {rope_deltas.shape} does not match prefill batch size {B}"
+            )
+        return rope_deltas
+
+
+class BatchGenerator:
+    """
+    Continuous batching with separate prompt processing and generation phases.
+
+    next() returns (prompt_responses, generation_responses) where:
+    - prompt_responses contains completed prompt-batch timing stats
+    - generation_responses is a list of GenerationBatch.Response objects
+    """
+
+    def __init__(
+        self,
+        model,
+        processor,
+        *,
+        max_tokens: int = DEFAULT_MAX_TOKENS,
+        stop_tokens: Optional[set] = None,
+        sampler: Optional[Callable[[mx.array], mx.array]] = None,
+        completion_batch_size: int = DEFAULT_COMPLETION_BATCH_SIZE,
+        prefill_batch_size: int = DEFAULT_PREFILL_BATCH_SIZE,
+        prefill_step_size: Optional[int] = DEFAULT_PREFILL_STEP_SIZE,
+        prompt_cache=None,
+        kv_bits=None,
+        kv_group_size: int = DEFAULT_KV_GROUP_SIZE,
+        kv_quant_scheme: str = DEFAULT_KV_QUANT_SCHEME,
+        quantized_kv_start: int = DEFAULT_QUANTIZED_KV_START,
+        compute_logprobs: bool = True,
+        top_logprobs_k: int = 0,
+        logits_processors: Optional[
+            List[Callable[[mx.array, mx.array], mx.array]]
+        ] = None,
+        stream=None,
+        apc_manager: Optional["_apc.APCManager"] = None,
+    ):
+        self.model = model
+        self.max_tokens = max_tokens
+        self.processor = processor
+        self.kv_bits = kv_bits
+        self.kv_group_size = kv_group_size
+        self.kv_quant_scheme = kv_quant_scheme
+        self.quantized_kv_start = quantized_kv_start
+        self.compute_logprobs = compute_logprobs
+        self.top_logprobs_k = top_logprobs_k
+        self.logits_processors = logits_processors or []
+        # APC: opt-out for KV-quantized caches. Plain KV models use block APC;
+        # mixed/custom cache models use exact prompt-cache snapshots.
+        self.apc_mode = None
+        if apc_manager is not None and kv_bits is not None:
+            apc_manager = None
+        if apc_manager is not None:
+            self.apc_mode = _apc.model_apc_mode(model)
+            if self.apc_mode is None:
+                apc_manager = None
+        self.apc_manager = apc_manager
+        self.tokenizer = (
+            processor.tokenizer if hasattr(processor, "tokenizer") else processor
+        )
+        self.sampler = sampler or (lambda x: mx.argmax(x, axis=-1))
+        self.uid_count = 0
+        self.prefill_step_size = prefill_step_size
+        self.prefill_batch_size = prefill_batch_size
+        self.completion_batch_size = completion_batch_size
+
+        self._stream = stream or generation_stream
+
+        self.tokenizer.stopping_criteria.add_eos_token_ids(stop_tokens)
+
+        self._generation_batch = GenerationBatch.empty(
+            self.model,
+            self.sampler,
+            self.tokenizer.stopping_criteria,
+            compute_logprobs=self.compute_logprobs,
+            top_logprobs_k=self.top_logprobs_k,
+        )
+        self._prompt_batch: Optional[PromptProcessingBatch] = None
+        self._unprocessed_sequences = []
+
+        self._prompt_tokens_counter = 0
+        self._prompt_time_counter = 0
+        self._gen_tokens_counter = 0
+        self._steps_counter = 0
+
+        self._wire_stack = contextlib.ExitStack()
+        self._wire_stack.enter_context(wired_limit(model, [self._stream]))
+
+    # ---------------- APC integration helpers ----------------
+    # Keys that are APC-only metadata; stripped from ``prompt_kwargs`` before
+    # the merged kwargs are passed to the language model forward.
+    _APC_PRIVATE_KEYS = APC_PRIVATE_PROMPT_KEYS
+
+    def _apc_extra_hash(self, prompt_kwargs: dict) -> int:
+        """Salt for the APC hash chain."""
+        if self.apc_manager is None:
+            return 0
+        if prompt_kwargs is None:
+            prompt_kwargs = {}
+        img = prompt_kwargs.get("_apc_image_hash")
+        if img is None:
+            pixel_values = prompt_kwargs.get("pixel_values")
+            img = _apc.hash_image_payload(pixel_values=pixel_values, image_ref=None)
+        tenant = prompt_kwargs.get("_apc_tenant")
+        return _apc.tenant_scoped_hash(tenant, img)
+
+    def _apc_pick_for(self, sequence) -> Optional[dict]:
+        """Look up an APC prefix for ``sequence``. Returns dict with matched
+        blocks + suffix metadata when there is a usable hit, else None.
+        """
+        if self.apc_manager is None:
+            return None
+        uid, ids_list, max_toks, prompt_kwargs, lps = sequence
+        if not ids_list or len(ids_list) < 2:
+            return None
+        # v1/v2: don't trim a prefix that contains image tokens — re-running
+        # vision merging on the suffix is the cheap path here.
+        image_token_id = getattr(self.model.config, "image_token_id", None) or getattr(
+            self.model.config, "image_token_index", None
+        )
+        extra_hash = self._apc_extra_hash(prompt_kwargs or {})
+        apc_mode = getattr(self, "apc_mode", "block")
+        if apc_mode == "exact":
+            exact_cache, exact_prefix_len = self.apc_manager.lookup_exact_cache(
+                ids_list,
+                extra_hash=extra_hash,
+            )
+            if (
+                exact_cache is not None
+                and exact_prefix_len > 0
+                and exact_prefix_len < len(ids_list)
+            ):
+                if (
+                    image_token_id is not None
+                    and image_token_id in ids_list[:exact_prefix_len]
+                ):
+                    return None
+                return {
+                    "matched_blocks": [],
+                    "warm_cache": exact_cache,
+                    "prefix_len": exact_prefix_len,
+                    "extra_hash": extra_hash,
+                    "full_input_ids": list(ids_list),
+                }
+            return None
+        matched, prefix_len = self.apc_manager.lookup_prefix(
+            ids_list, extra_hash=extra_hash
+        )
+        exact_cache = None
+        exact_prefix_len = 0
+        if prefix_len < len(ids_list):
+            exact_cache, exact_prefix_len = self.apc_manager.lookup_exact_cache(
+                ids_list,
+                extra_hash=extra_hash,
+                min_prefix_tokens=prefix_len,
+            )
+        warm_cache = None
+        disk_prefix_len = 0
+        if max(prefix_len, exact_prefix_len) < len(ids_list):
+            warm_cache, disk_prefix_len = self.apc_manager.lookup_prefix_disk_cache(
+                ids_list,
+                extra_hash=extra_hash,
+                min_prefix_tokens=max(prefix_len, exact_prefix_len),
+                allow_memory_overlap=max(prefix_len, exact_prefix_len) > 0,
+            )
+        if disk_prefix_len > max(
+            prefix_len, exact_prefix_len
+        ) and disk_prefix_len < len(ids_list):
+            if matched:
+                self.apc_manager.release(matched)
+            if (
+                image_token_id is not None
+                and image_token_id in ids_list[:disk_prefix_len]
+            ):
+                return None
+            return {
+                "matched_blocks": [],
+                "warm_cache": warm_cache,
+                "prefix_len": disk_prefix_len,
+                "extra_hash": extra_hash,
+                "full_input_ids": list(ids_list),
+            }
+        if exact_prefix_len > prefix_len and exact_prefix_len < len(ids_list):
+            if matched:
+                self.apc_manager.release(matched)
+            if (
+                image_token_id is not None
+                and image_token_id in ids_list[:exact_prefix_len]
+            ):
+                return None
+            return {
+                "matched_blocks": [],
+                "warm_cache": exact_cache,
+                "prefix_len": exact_prefix_len,
+                "extra_hash": extra_hash,
+                "full_input_ids": list(ids_list),
+            }
+        if prefix_len > 0 and prefix_len < len(ids_list):
+            if image_token_id is not None and image_token_id in ids_list[:prefix_len]:
+                self.apc_manager.release(matched)
+                return None
+            return {
+                "matched_blocks": matched,
+                "prefix_len": prefix_len,
+                "extra_hash": extra_hash,
+                "full_input_ids": list(ids_list),
+            }
+        if matched:
+            self.apc_manager.release(matched)
+        return None
+
+    def _build_mixed_prompt_batch(
+        self, sequences: List[tuple]
+    ) -> Optional["PromptProcessingBatch"]:
+        """Build a multi-row PromptProcessingBatch admitting ``sequences``.
+
+        Each row is independently looked up in APC. Warm rows have their
+        suffixes prefilled against pre-populated K/V; cold rows prefill from
+        scratch in the same batch. Right-padding aligns RoPE positions
+        across rows with different prefix/suffix lengths.
+
+        Returns ``None`` if APC is disabled (in which case the caller should
+        use the cold-only path).
+        """
+        if self.apc_manager is None:
+            return None
+
+        picks: List[Optional[dict]] = [self._apc_pick_for(s) for s in sequences]
+        any_warm = any(p is not None for p in picks)
+        if not any_warm:
+            return None  # caller falls back to cold-only path
+
+        uids = [s[0] for s in sequences]
+        full_ids = [list(s[1]) for s in sequences]
+        max_tokens_list = [s[2] for s in sequences]
+        prompt_kwargs_list = [s[3] for s in sequences]
+        logits_processors = [s[4] for s in sequences]
+
+        # Per-row prefix length and suffix tokens
+        prefix_lens = [p["prefix_len"] if p else 0 for p in picks]
+        suffix_ids_list = [full_ids[i][prefix_lens[i] :] for i in range(len(sequences))]
+        suffix_lens = [len(s) for s in suffix_ids_list]
+
+        max_suffix_len = max(suffix_lens)
+        right_pad_per_row = [max_suffix_len - s for s in suffix_lens]
+
+        # Source inputs_embeds: every row's prompt_kwargs holds the full-prompt
+        # embeddings. Slice to suffix per-row, right-pad to max_suffix_len, stack.
+        suffix_embeds_per_row: List[mx.array] = []
+        for i, kw in enumerate(prompt_kwargs_list):
+            if kw is None or kw.get("inputs_embeds") is None:
+                raise ValueError("APC mixed prefill requires precomputed inputs_embeds")
+            full = kw["inputs_embeds"]  # [1, full_len, D]
+            suff = full[:, prefix_lens[i] :, :]
+            pad = right_pad_per_row[i]
+            if pad > 0:
+                pad_emb = mx.zeros(
+                    (suff.shape[0], pad, suff.shape[-1]), dtype=suff.dtype
+                )
+                suff = mx.concatenate([suff, pad_emb], axis=1)
+            suffix_embeds_per_row.append(suff)
+        inputs_embeds = mx.concatenate(suffix_embeds_per_row, axis=0)
+
+        # Merge prompt-side kwargs (excluding inputs_embeds, which we've just
+        # rebuilt). Per-batch tensors get concatenated across rows; scalars
+        # take the first row's value (matches the existing cold-only path).
+        # APC-private keys (e.g. tenant salt) are dropped — they're consumed
+        # in _apc_extra_hash, never forwarded to the model.
+        merged_kwargs: dict = {}
+        per_row_keys: dict = {}
+        batch_size = len(prompt_kwargs_list)
+        for i, kw in enumerate(prompt_kwargs_list):
+            if not kw:
+                continue
+            full_len = len(full_ids[i])
+            prefix_len = prefix_lens[i]
+            right_pad = right_pad_per_row[i]
+            for k, v in kw.items():
+                if k == "inputs_embeds" or k in self._APC_PRIVATE_KEYS:
+                    continue
+                if isinstance(v, mx.array) and v.ndim > 0 and v.shape[0] >= 1:
+                    row_v = _prompt_kwarg_row(v, i, batch_size)
+                    if _is_sequence_aligned_prompt_kwarg(k, row_v, full_len):
+                        row_v = row_v[:, prefix_len:, ...]
+                        row_v = _pad_sequence_aligned_prompt_kwarg(
+                            row_v,
+                            max_suffix_len,
+                            left=False,
+                        )
+                    per_row_keys.setdefault(k, []).append(row_v)
+                elif k not in merged_kwargs:
+                    merged_kwargs[k] = v
+        for k, vs in per_row_keys.items():
+            merged_kwargs[k] = mx.concatenate(vs, axis=0)
+
+        apc_mode = getattr(self, "apc_mode", "block")
+        if apc_mode == "exact":
+            row_caches = [
+                p["warm_cache"] if p is not None else self.model.make_cache()
+                for p in picks
+            ]
+            warm_cache, _ = _apc.make_warm_batch_exact_cache_multi(
+                row_caches,
+                prefix_lens,
+            )
+            if warm_cache is None:
+                return None
+        else:
+            # Build the multi-row warm cache (zeros for cold rows, K/V for warm).
+            num_layers = (
+                len(self.model.make_cache())
+                if hasattr(self.model, "make_cache")
+                else len(self.model.layers)
+            )
+            warm_cache, _ = _apc.make_warm_batch_kv_cache_multi(
+                picks, num_layers=num_layers
+            )
+
+        apc_meta = [
+            {
+                "full_input_ids": full_ids[i],
+                "prefix_len": prefix_lens[i],
+                "extra_hash": (
+                    picks[i]["extra_hash"]
+                    if picks[i]
+                    else self._apc_extra_hash(prompt_kwargs_list[i] or {})
+                ),
+                "apc_blocks": picks[i].get("matched_blocks", []) if picks[i] else [],
+                "checkpoint_len": (
+                    max(
+                        1,
+                        len(full_ids[i]) - self.apc_manager.exact_cache_guard_tokens,
+                    )
+                    if apc_mode == "exact"
+                    else 0
+                ),
+            }
+            for i in range(len(sequences))
+        ]
+
+        prompt_batch_cls = _generate_module_override(
+            "PromptProcessingBatch", PromptProcessingBatch
+        )
+        return prompt_batch_cls(
+            model=self.model,
+            uids=uids,
+            input_ids=suffix_ids_list,
+            max_tokens=max_tokens_list,
+            inputs_embeds=inputs_embeds,
+            prompt_kwargs=merged_kwargs,
+            logits_processors=logits_processors,
+            prefill_step_size=self.prefill_step_size,
+            kv_bits=self.kv_bits,
+            kv_group_size=self.kv_group_size,
+            kv_quant_scheme=self.kv_quant_scheme,
+            warm_cache=warm_cache,
+            apc_meta=apc_meta,
+            apc_manager=self.apc_manager,
+            right_pad_per_row=right_pad_per_row,
+            suffix_lens=suffix_lens,
+            apc_mode=apc_mode,
+        )
+
+    def _build_apc_meta_for_cold(
+        self,
+        input_ids_list: List[List[int]],
+        prompt_kwargs_list: List[Optional[dict]],
+    ) -> Optional[List[Optional[dict]]]:
+        """Build per-row harvest metadata for a cold-prefill batch so the
+        produced K/V are added to APC after prefill.
+        """
+        if self.apc_manager is None:
+            return None
+        meta: List[Optional[dict]] = []
+        for ids_list, kw in zip(input_ids_list, prompt_kwargs_list):
+            extra_hash = self._apc_extra_hash(kw or {})
+            meta.append(
+                {
+                    "full_input_ids": list(ids_list),
+                    "prefix_len": 0,
+                    "extra_hash": extra_hash,
+                    "apc_blocks": [],
+                    "checkpoint_len": (
+                        max(
+                            1,
+                            len(ids_list) - self.apc_manager.exact_cache_guard_tokens,
+                        )
+                        if getattr(self, "apc_mode", "block") == "exact"
+                        else 0
+                    ),
+                }
+            )
+        return meta
+
+    @property
+    def stream(self):
+        return self._stream
+
+    def close(self):
+        if self._wire_stack is not None:
+            self._wire_stack.close()
+            self._wire_stack = None
+
+    def __del__(self):
+        self.close()
+
+    def insert(
+        self,
+        prompts,
+        max_tokens: Union[List[int], int, None] = None,
+        prompt_kwargs: Optional[List[dict]] = None,
+        logits_processors: Optional[
+            List[Optional[List[Callable[[mx.array, mx.array], mx.array]]]]
+        ] = None,
+    ):
+        uids = []
+
+        if max_tokens is None or isinstance(max_tokens, int):
+            max_tokens = [max_tokens or self.max_tokens] * len(prompts)
+
+        if prompt_kwargs is None:
+            prompt_kwargs = [{}] * len(prompts)
+        if logits_processors is None:
+            logits_processors = [self.logits_processors] * len(prompts)
+        elif len(logits_processors) != len(prompts):
+            raise ValueError("Insufficient number of logits_processors provided")
+
+        for p, m, kw, lp in zip(prompts, max_tokens, prompt_kwargs, logits_processors):
+            self._unprocessed_sequences.append((self.uid_count, p, m, kw, lp))
+            uids.append(self.uid_count)
+            self.uid_count += 1
+        # Sort in ascending order of length
+        self._unprocessed_sequences = sorted(
+            self._unprocessed_sequences, key=lambda x: len(x[1])
+        )
+        return uids
+
+    def remove(self, uid) -> bool:
+        """Remove a sequence from the batch by uid."""
+        with mx.stream(self._stream):
+            # Waiting in the queue.
+            for i, (seq_uid, _, _, _, _) in enumerate(self._unprocessed_sequences):
+                if seq_uid == uid:
+                    self._unprocessed_sequences.pop(i)
+                    return True
+
+            # Being prefilled
+            if self._prompt_batch is not None and uid in self._prompt_batch.uids:
+                if len(self._prompt_batch.uids) == 1:
+                    self._prompt_batch.uids = []
+                    self._prompt_batch.prompt_cache = []
+                    self._prompt_batch = None
+                    mx.clear_cache()
+                    return True
+
+            # Already decoding.
+            if uid in self._generation_batch.uids:
+                idx = self._generation_batch.uids.index(uid)
+                keep = [i for i in range(len(self._generation_batch.uids)) if i != idx]
+                self._generation_batch.filter(keep)
+                return True
+
+            return False
+
+    @property
+    def unprocessed_prompts(self):
+        """Backward-compatible alias for server flush logic."""
+        return self._unprocessed_sequences
+
+    @property
+    def has_pending_prompts(self):
+        """True if there are prompts waiting or being processed."""
+        return len(self._unprocessed_sequences) > 0 or self._prompt_batch is not None
+
+    @property
+    def has_work(self):
+        """True if there is any remaining work."""
+        return (
+            len(self._generation_batch) > 0
+            or self._prompt_batch is not None
+            or len(self._unprocessed_sequences) > 0
+        )
+
+    def stats(self):
+        """Return accumulated batch statistics."""
+        stats = BatchStats()
+        stats.prompt_tokens = self._prompt_tokens_counter
+        stats.prompt_time = self._prompt_time_counter
+        stats.prompt_tps = (
+            self._prompt_tokens_counter / self._prompt_time_counter
+            if self._prompt_time_counter > 0
+            else 0
+        )
+        stats.generation_tokens = self._gen_tokens_counter
+        stats.peak_memory = mx.get_peak_memory() / 1e9
+        return stats
+
+    @staticmethod
+    def _record_prompt_batch_time(prompt_batch, elapsed_s: float) -> None:
+        recorder = getattr(prompt_batch, "record_prompt_time", None)
+        if callable(recorder):
+            recorder(elapsed_s)
+
+    @staticmethod
+    def _prompt_batch_progress(prompt_batch) -> List[PromptProgress]:
+        progress = getattr(prompt_batch, "prompt_progress", None)
+        if callable(progress):
+            return progress()
+        return []
+
+    def _next(self, **kwargs):
+        generation_responses = []
+        prompt_responses = []
+
+        # Decode-first: always emit a generation step before touching prefill.
+        if len(self._generation_batch) > 0:
+            generation_responses = self._generation_batch.next()
+            self._gen_tokens_counter += len(generation_responses)
+            self._steps_counter += 1
+            if self._steps_counter % 50 == 0:
+                mx.eval([c.state for c in self._generation_batch.prompt_cache])
+                mx.clear_cache()
+
+        if len(self._generation_batch) >= self.completion_batch_size:
+            return prompt_responses, generation_responses
+
+        if self._prompt_batch is not None:
+            if self._prompt_batch.needs_processing():
+                tic = time.perf_counter()
+                n = self._prompt_batch.prompt_step()
+                elapsed = time.perf_counter() - tic
+                self._prompt_time_counter += elapsed
+                self._record_prompt_batch_time(self._prompt_batch, elapsed)
+                self._prompt_tokens_counter += n
+                return prompt_responses, generation_responses
+
+            tic = time.perf_counter()
+            gen_batch = self._prompt_batch.generate(
+                self.sampler,
+                self.tokenizer.stopping_criteria,
+                compute_logprobs=self.compute_logprobs,
+                top_logprobs_k=self.top_logprobs_k,
+            )
+            elapsed = time.perf_counter() - tic
+            self._prompt_time_counter += elapsed
+            self._record_prompt_batch_time(self._prompt_batch, elapsed)
+            prompt_responses = self._prompt_batch_progress(self._prompt_batch)
+            self._generation_batch.extend(gen_batch)
+            self._prompt_batch = None
+            mx.clear_cache()
+            return prompt_responses, generation_responses
+
+        num_active = len(self._generation_batch)
+        num_to_add = self.completion_batch_size - num_active
+        if self._unprocessed_sequences and num_to_add >= self.prefill_batch_size:
+            # Take up to prefill_batch_size pending sequences. If APC is on
+            # and at least one of them has a prefix hit, build a mixed
+            # warm/cold PromptProcessingBatch with right-padded suffixes so
+            # warm and cold rows prefill in a single forward pass.
+            n = min(self.prefill_batch_size, len(self._unprocessed_sequences))
+            sequences = self._unprocessed_sequences[:n]
+            if logger.isEnabledFor(logging.DEBUG) and os.environ.get("APC_DEBUG"):
+                logger.warning(
+                    "APC admit n=%d (pending=%d)",
+                    n,
+                    len(self._unprocessed_sequences),
+                )
+            mixed = self._build_mixed_prompt_batch(sequences)
+            if mixed is not None:
+                self._unprocessed_sequences = self._unprocessed_sequences[n:]
+                self._prompt_batch = mixed
+                self._prompt_tokens_counter += self._prompt_batch.total_prompt_tokens
+                if self._prompt_batch.needs_processing():
+                    tic = time.perf_counter()
+                    nstep = self._prompt_batch.prompt_step()
+                    elapsed = time.perf_counter() - tic
+                    self._prompt_time_counter += elapsed
+                    self._record_prompt_batch_time(self._prompt_batch, elapsed)
+                else:
+                    tic = time.perf_counter()
+                    gen_batch = self._prompt_batch.generate(
+                        self.sampler,
+                        self.tokenizer.stopping_criteria,
+                        compute_logprobs=self.compute_logprobs,
+                        top_logprobs_k=self.top_logprobs_k,
+                    )
+                    elapsed = time.perf_counter() - tic
+                    self._prompt_time_counter += elapsed
+                    self._record_prompt_batch_time(self._prompt_batch, elapsed)
+                    prompt_responses = self._prompt_batch_progress(self._prompt_batch)
+                    self._generation_batch.extend(gen_batch)
+                    self._prompt_batch = None
+                    mx.clear_cache()
+                return prompt_responses, generation_responses
+
+            self._unprocessed_sequences = self._unprocessed_sequences[n:]
+
+            uids = [s[0] for s in sequences]
+            input_ids = [s[1] for s in sequences]
+            max_tokens_list = [s[2] for s in sequences]
+            prompt_kwargs_list = [s[3] for s in sequences]
+            logits_processors = [s[4] for s in sequences]
+
+            inputs_embeds, merged_kwargs = _merge_prefill_prompt_kwargs(
+                prompt_kwargs_list, input_ids
+            )
+
+            # APC: also harvest cold-prefill prefixes so future requests hit.
+            apc_meta = self._build_apc_meta_for_cold(input_ids, prompt_kwargs_list)
+
+            prompt_batch_cls = _generate_module_override(
+                "PromptProcessingBatch", PromptProcessingBatch
+            )
+            self._prompt_batch = prompt_batch_cls(
+                model=self.model,
+                uids=uids,
+                input_ids=input_ids,
+                max_tokens=max_tokens_list,
+                inputs_embeds=inputs_embeds,
+                prompt_kwargs=merged_kwargs,
+                logits_processors=logits_processors,
+                prefill_step_size=self.prefill_step_size,
+                kv_bits=self.kv_bits,
+                kv_group_size=self.kv_group_size,
+                kv_quant_scheme=self.kv_quant_scheme,
+                apc_meta=apc_meta,
+                apc_manager=self.apc_manager,
+                apc_mode=self.apc_mode,
+            )
+            self._prompt_tokens_counter += self._prompt_batch.total_prompt_tokens
+
+            if self._prompt_batch.needs_processing():
+                tic = time.perf_counter()
+                n = self._prompt_batch.prompt_step()
+                elapsed = time.perf_counter() - tic
+                self._prompt_time_counter += elapsed
+                self._record_prompt_batch_time(self._prompt_batch, elapsed)
+            else:
+                tic = time.perf_counter()
+                gen_batch = self._prompt_batch.generate(
+                    self.sampler,
+                    self.tokenizer.stopping_criteria,
+                    compute_logprobs=self.compute_logprobs,
+                    top_logprobs_k=self.top_logprobs_k,
+                )
+                elapsed = time.perf_counter() - tic
+                self._prompt_time_counter += elapsed
+                self._record_prompt_batch_time(self._prompt_batch, elapsed)
+                prompt_responses = self._prompt_batch_progress(self._prompt_batch)
+                self._generation_batch.extend(gen_batch)
+                self._prompt_batch = None
+                mx.clear_cache()
+
+            return prompt_responses, generation_responses
+
+        return prompt_responses, generation_responses
+
+    def next(self, **kwargs):
+        with mx.stream(self._stream):
+            return self._next(**kwargs)
+
+
+def batch_generate(
+    model,
+    processor,
+    images: Union[str, List[str]] = None,
+    audios: Union[str, List[str]] = None,
+    prompts: List[str] = None,
+    max_tokens: Union[int, List[int]] = 128,
+    verbose: bool = False,
+    group_by_shape: bool = True,
+    track_image_sizes: bool = True,
+    **kwargs,
+):
+    """
+    Generate responses for the given batch of prompts with variable-sized images.
+
+    This function implements the transformers-style approach to batching:
+    1. Group images with the same shape for efficient batch processing
+    2. Process each group as a batch (no padding waste within groups)
+    3. Track original image sizes for proper attention masking
+    4. Restore results to original batch order
+
+    Key insight: Instead of padding all images to the same spatial dimensions
+    (which wastes computation and may hurt accuracy), we group same-sized
+    images together so there's zero padding within each group.
+
+    Args:
+       model (nn.Module): The language model.
+       processor (PreTrainedTokenizer): The tokenizer/processor.
+       images (Union[str, List[str]]): Images (paths, URLs, or PIL images).
+       audios (Union[str, List[str]]): Audio files (not yet supported for batching).
+       prompts (List[str]): The input prompts.
+       max_tokens (Union[int, List[int]]): Maximum number of output tokens. This
+          can be per prompt if a list is provided.
+       verbose (bool): If ``True``, print tokens and timing information.
+       group_by_shape (bool): If ``True``, group same-shaped images for efficient
+          batch processing.
+       track_image_sizes (bool): If ``True``, track and return original image sizes.
+       kwargs: The remaining options get passed to :obj:`BatchGenerator`.
+          See :obj:`BatchGenerator` for more details.
+
+    Returns:
+        BatchResponse with generated texts, statistics, and optionally image_sizes.
+    """
+    from PIL import Image
+
+    from ..utils import process_image
+
+    processor.detokenizer.reset()
+    tokenizer = processor.tokenizer if hasattr(processor, "tokenizer") else processor
+
+    # Handle single image case
+    if isinstance(images, str):
+        images = [images]
+
+    # Handle no images case
+    if images is None:
+        texts, stats = _generate_batch(
+            model, processor, prompts, None, max_tokens, verbose, **kwargs
+        )
+        return BatchResponse(texts, stats)
+
+    # Load and preprocess images
+    image_processor = (
+        processor.image_processor if hasattr(processor, "image_processor") else None
+    )
+
+    processed_images = []
+    image_sizes_original = []
+    for img in images:
+        if isinstance(img, str):
+            pil_img = process_image(img, None, image_processor)
+        elif isinstance(img, Image.Image):
+            pil_img = img
+        else:
+            pil_img = img
+        processed_images.append(pil_img)
+        # Track original size
+        if hasattr(pil_img, "height"):
+            image_sizes_original.append((pil_img.height, pil_img.width))
+        else:
+            image_sizes_original.append((0, 0))
+
+    # Group images by shape for efficient processing (no padding within groups)
+    if group_by_shape and len(processed_images) > 1:
+        grouped_images, grouped_indices = group_images_by_shape(processed_images)
+
+        if verbose:
+            print(f"[batch_generate] Found {len(grouped_images)} unique image shapes")
+    else:
+        # Single image or grouping disabled - treat as one group
+        shape = (
+            (processed_images[0].height, processed_images[0].width)
+            if processed_images
+            else (0, 0)
+        )
+        grouped_images = {shape: processed_images}
+        grouped_indices = {shape: list(range(len(processed_images)))}
+
+    # Process each shape group
+    all_texts = [None] * len(prompts)
+    all_image_sizes = [None] * len(prompts)
+    total_stats = BatchStats()
+
+    for shape, indices in grouped_indices.items():
+        # Get images and prompts for this shape group
+        group_images = [processed_images[i] for i in indices]
+        group_prompts = [prompts[i] for i in indices]
+        group_sizes = [image_sizes_original[i] for i in indices]
+
+        # Handle per-sample max_tokens
+        if isinstance(max_tokens, list):
+            group_max_tokens = [max_tokens[i] for i in indices]
+        else:
+            group_max_tokens = max_tokens
+
+        group_kwargs = dict(kwargs)
+        logits_processors = group_kwargs.get("logits_processors")
+        if logits_processors is not None and isinstance(logits_processors, list):
+            if not logits_processors or all(callable(p) for p in logits_processors):
+                group_kwargs["logits_processors"] = logits_processors
+            else:
+                group_kwargs["logits_processors"] = [
+                    logits_processors[i] for i in indices
+                ]
+
+        # Process the entire group at once (same shape = no padding needed)
+        chunk_texts, chunk_stats = _generate_batch(
+            model,
+            processor,
+            group_prompts,
+            group_images,
+            group_max_tokens,
+            **group_kwargs,
+        )
+
+        # Store results in original order
+        for j, orig_idx in enumerate(indices):
+            all_texts[orig_idx] = chunk_texts[j]
+            all_image_sizes[orig_idx] = group_sizes[j]
+
+        # Accumulate stats
+        total_stats.prompt_tokens += chunk_stats.prompt_tokens
+        total_stats.prompt_time += chunk_stats.prompt_time
+        total_stats.generation_tokens += chunk_stats.generation_tokens
+        total_stats.generation_time += chunk_stats.generation_time
+
+    mx.clear_cache()
+
+    # Compute final stats
+    if total_stats.prompt_time > 0:
+        total_stats.prompt_tps = total_stats.prompt_tokens / total_stats.prompt_time
+    if total_stats.generation_time > 0:
+        total_stats.generation_tps = (
+            total_stats.generation_tokens / total_stats.generation_time
+        )
+    total_stats.peak_memory = mx.get_peak_memory() / 1e9
+
+    if verbose:
+        print(f"[batch_generate] Finished processing {len(prompts)} samples")
+        print(
+            f"[batch_generate] Prompt: {total_stats.prompt_tokens} tokens, {total_stats.prompt_tps:.3f} tokens-per-sec"
+        )
+        print(
+            f"[batch_generate] Generation: {total_stats.generation_tokens} tokens, "
+            f"{total_stats.generation_tps:.3f} tokens-per-sec"
+        )
+        print(f"[batch_generate] Peak memory: {total_stats.peak_memory:.3f} GB")
+
+    response = BatchResponse(all_texts, total_stats)
+    if track_image_sizes:
+        response.image_sizes = all_image_sizes
+    return response
+
+
+def _clone_or_share_logits_processor(processor):
+    if hasattr(processor, "clone"):
+        return processor.clone()
+    warnings.warn(
+        "Sharing logits processor across batch entries because it does not "
+        "implement clone(). Stateful logits processors should implement clone() "
+        "to avoid shared state across sequences.",
+        RuntimeWarning,
+        stacklevel=2,
+    )
+    return processor
+
+
+def _generate_batch(
+    model,
+    processor,
+    prompts: List[str],
+    images: List = None,
+    max_tokens: Union[int, List[int]] = 100,
+    verbose: bool = False,
+    **kwargs,
+) -> Tuple[List[str], BatchStats]:
+
+    tokenizer = processor.tokenizer if hasattr(processor, "tokenizer") else processor
+    batch_size = len(prompts)
+    logits_processors = kwargs.pop("logits_processors", None)
+
+    num_images_list = [
+        1 if i < (len(images) if images is not None else 0) else 0
+        for i in range(len(prompts))
+    ]
+    formatted_prompts = [
+        apply_chat_template(
+            processor,
+            model.config,
+            p,
+            num_images=num_images_list[i],
+        )
+        for i, p in enumerate(prompts)
+    ]
+
+    add_special_tokens = (
+        getattr(processor, "chat_template", None) is None
+        if model.config.model_type in ["gemma3", "gemma3n", "gemma4"]
+        else True
+    )
+
+    resize_shape = normalize_resize_shape(kwargs.pop("resize_shape", None))
+    image_token_index = getattr(model.config, "image_token_index", None)
+
+    inputs = prepare_inputs(
+        processor,
+        images=images,
+        audio=None,
+        prompts=formatted_prompts,
+        image_token_index=image_token_index,
+        resize_shape=resize_shape,
+        add_special_tokens=add_special_tokens,
+        pad_to_uniform_size=False,  # Since images are pre-grouped by shape, they're already uniform size
+    )
+    input_ids = inputs.get("input_ids", None)
+    pixel_values = inputs.get("pixel_values", None)
+    mask = inputs.get("attention_mask", None)
+
+    data_kwargs = {
+        k: v
+        for k, v in inputs.items()
+        if k not in ["input_ids", "pixel_values", "attention_mask"]
+    }
+
+    if getattr(model, "no_chunked_prefill", False):
+        kwargs.pop("prefill_step_size", None)
+        kwargs["prefill_step_size"] = None
+
+    # Use batch_size for prefill and completion to ensure consistent processing
+    gen = BatchGenerator(
+        model.language_model,
+        processor,
+        prefill_batch_size=batch_size,
+        completion_batch_size=batch_size,
+        compute_logprobs=False,
+        **kwargs,
+    )
+
+    embedding_output = model.get_input_embeddings(
+        input_ids, pixel_values, mask=mask, **data_kwargs
+    )
+
+    gen_kwargs = {**data_kwargs, **embedding_output.to_dict()}
+
+    if logits_processors and all(
+        callable(processor) for processor in logits_processors
+    ):
+        logits_processors = [
+            [_clone_or_share_logits_processor(p) for p in logits_processors]
+            for _ in range(batch_size)
+        ]
+
+    uids = gen.insert(
+        input_ids.tolist(),
+        max_tokens,
+        prompt_kwargs=_split_prompt_kwargs_per_row(gen_kwargs, batch_size),
+        logits_processors=logits_processors,
+    )
+    results = {uid: [] for uid in uids}
+
+    tic = time.perf_counter()
+    while gen.has_work:
+        _, generation_responses = gen.next()
+        for r in generation_responses:
+            if r.finish_reason != "stop":
+                results[r.uid].append(r.token)
+    total_time = time.perf_counter() - tic
+
+    gen.close()
+
+    detokenizer = processor.detokenizer
+    texts = []
+    for uid in uids:
+        detokenizer.reset()
+        for t in results[uid]:
+            detokenizer.add_token(t)
+        detokenizer.finalize()
+        texts.append(detokenizer.text)
+
+    stats = gen.stats()
+    stats.generation_time = total_time - stats.prompt_time
+    if stats.generation_time > 0:
+        stats.generation_tps = stats.generation_tokens / stats.generation_time
+    return texts, stats
